requirement.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework-jwt==1.11.0
mysqlclient==2.2.7
PyJWT==1.7.1
PyMySQL==1.1.2
sqlparse==0.5.3
